Move availableTodoGroups to a function
